+++ demo/src/main/java/ProIntermodular/demo/repository/IUsuariosRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IUsuariosRepository extends JpaRepository<Usuarios,Long> {
+public interface IUsuariosRepository extends JpaRepository<Usuarios,Integer> {
     Usuarios findByEmailAndContrasena(String email, String contrasena);
 // Metodo para buscar al usuario por correo
     Optional<Usuarios> findByEmail(String email);
